+++ R/translate.R
@@ -1,7 +1,3 @@
-obj_maybe_translate_encoding <- function(x) {
-  .Call(vctrs_maybe_translate_encoding, x)
-}
-
-obj_maybe_translate_encoding2 <- function(x, y) {
-  .Call(vctrs_maybe_translate_encoding2, x, y)
+vec_normalize_encoding <- function(x) {
+  .Call(vctrs_normalize_encoding, x)
 }