+++ src/vctrs.h
@@ -584,11 +584,6 @@ SEXP vec_posixlt_restore(SEXP x, SEXP to, const enum vctrs_owned owned);
 SEXP date_datetime_ptype2(SEXP x, SEXP y);
 SEXP datetime_datetime_ptype2(SEXP x, SEXP y);
 
-// Character translation ---------------------------------------------
-
-SEXP obj_maybe_translate_encoding(SEXP x, r_ssize size);
-SEXP obj_maybe_translate_encoding2(SEXP x, r_ssize x_size, SEXP y, r_ssize y_size);
-
 // Growable vector ----------------------------------------------
 
 struct growable {